refactor(spf)!: name the SPF Mux adapters for their flavors
The two adapters were `mux` and `mux-audio-only` — the video one unqualified,
the audio one named for the composition underneath rather than for what it
plays. They are now `mux-video` and `mux-audio`, matching the elements,
components, and sandbox presets that consume them.

The Medias follow: `MuxMedia` becomes `MuxVideoMedia` and `MuxAudioOnlyMedia`
becomes `MuxAudioMedia`. That gives up sharing a class name with the
hls.js-backed `MuxMedia` — worth it, since the mirror could only ever hold for
one of the two flavors, and the audio one needed a distinct name regardless.

What both flavors share keeps the unqualified name: `MuxMediaMixin`,
`MuxMediaProps`, `MuxMediaAPI`, and `muxMediaDefaultProps` are Mux identity, not
either flavor's.

`mux-audio` still imports the mixin from `../mux-video/adapter` rather than the
barrel, which would drag in the full HLS engine — the same shape as
`simple-hls-audio-only` importing its error surface from `../simple-hls`.

BREAKING CHANGE: `@videojs/spf/mux` is now `@videojs/spf/mux-video`, and
`@videojs/spf/mux-audio-only` is now `@videojs/spf/mux-audio`.
