+++ packages/spf/package.json
@@ -43,15 +43,15 @@
       "development": "./dist/dev/background-video.js",
       "default": "./dist/default/background-video.js"
     },
-    "./mux": {
-      "types": "./dist/dev/mux.d.ts",
-      "development": "./dist/dev/mux.js",
-      "default": "./dist/default/mux.js"
+    "./mux-audio": {
+      "types": "./dist/dev/mux-audio.d.ts",
+      "development": "./dist/dev/mux-audio.js",
+      "default": "./dist/default/mux-audio.js"
     },
-    "./mux-audio-only": {
-      "types": "./dist/dev/mux-audio-only.d.ts",
-      "development": "./dist/dev/mux-audio-only.js",
-      "default": "./dist/default/mux-audio-only.js"
+    "./mux-video": {
+      "types": "./dist/dev/mux-video.d.ts",
+      "development": "./dist/dev/mux-video.js",
+      "default": "./dist/default/mux-video.js"
     },
     "./simple-hls": {
       "types": "./dist/dev/simple-hls.d.ts",
