+++ packages/spf/src/playback/adapters/mux-video/media.ts
@@ -1,14 +1,15 @@
 import { SimpleHlsMedia } from '../simple-hls/media';
 import { MuxMediaMixin } from './adapter';
 
-const MuxMediaBase = MuxMediaMixin(SimpleHlsMedia);
+const MuxVideoMediaBase = MuxMediaMixin(SimpleHlsMedia);
 
 /**
  * The Mux Media over the SPF HLS engine.
  *
  * Mirrors `@videojs/media/dom/mux`'s hls.js-backed `MuxMedia` — same class shape,
  * same `src`/`source` relationship, same derived `contentData` — over a different
- * engine. It carries no `engine` or `preferPlayback`: SPF publishes no
+ * engine, though named for the flavor rather than sharing that class's name. It
+ * carries no `engine` or `preferPlayback`: SPF publishes no
  * engine-shaped config for a consumer to pass, so the source is Mux identity and
  * nothing else.
  *
@@ -16,4 +17,4 @@ const MuxMediaBase = MuxMediaMixin(SimpleHlsMedia);
  * unsupported DRM, and `alternativeMediaSuggestion` points at the hls.js-backed
  * import, so a protected source fails with copy that says where to go.
  */
-export class MuxMedia extends MuxMediaBase {}
+export class MuxVideoMedia extends MuxVideoMediaBase {}
