+++ packages/spf/src/playback/adapters/mux-video/tests/media.test.ts
@@ -1,5 +1,5 @@
 /**
- * SPF-backed MuxMedia tests.
+ * SPF-backed MuxVideoMedia tests.
  *
  * Mirrors the coverage of the hls.js-backed `MuxMedia`
  * (`packages/media/src/dom/mux/tests/mux-media.test.ts`), minus everything that
@@ -7,36 +7,36 @@
  * identity and nothing else.
  */
 import { describe, expect, it, vi } from 'vitest';
-import { MuxMedia } from '../media';
+import { MuxVideoMedia } from '../media';
 
-describe('MuxMedia', () => {
+describe('MuxVideoMedia', () => {
   it('defaults source to null', () => {
-    expect(new MuxMedia().source).toBe(null);
+    expect(new MuxVideoMedia().source).toBe(null);
   });
 
   it('derives src from source.playbackId', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     media.source = { playbackId: 'abc123' };
 
     expect(media.src).toBe('https://stream.mux.com/abc123.m3u8');
   });
 
   it('derives src using the custom domain', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     media.source = { playbackId: 'abc123', customDomain: 'video.example.com' };
 
     expect(media.src).toBe('https://stream.video.example.com/abc123.m3u8');
   });
 
   it('appends playback params as snake_case query params', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     media.source = { playbackId: 'abc123', playback: { maxResolution: '720p' } };
 
     expect(media.src).toBe('https://stream.mux.com/abc123.m3u8?max_resolution=720p');
   });
 
   it('clears src when source is cleared', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     media.source = { playbackId: 'abc123' };
     media.source = null;
 
@@ -44,14 +44,14 @@ describe('MuxMedia', () => {
   });
 
   it('parses source from a Mux stream src', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     media.src = 'https://stream.mux.com/abc123.m3u8';
 
     expect(media.source).toEqual({ playbackId: 'abc123' });
   });
 
   it('parses the custom domain and playback params from a Mux stream src', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     media.src = 'https://stream.video.example.com/abc123.m3u8?max_resolution=720p';
 
     expect(media.source).toEqual({
@@ -62,7 +62,7 @@ describe('MuxMedia', () => {
   });
 
   it('keeps a non-Mux src as a plain source url', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     media.src = 'https://example.com/stream.m3u8';
 
     expect(media.source).toEqual({ src: 'https://example.com/stream.m3u8' });
@@ -70,14 +70,14 @@ describe('MuxMedia', () => {
   });
 
   it('plays a non-Mux source url given through source', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     media.source = { src: 'https://example.com/stream.m3u8' };
 
     expect(media.src).toBe('https://example.com/stream.m3u8');
   });
 
   it('exposes the content poster and storyboard derived from source', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     media.source = { playbackId: 'abc123' };
 
     expect(media.contentData).toEqual({
@@ -87,7 +87,7 @@ describe('MuxMedia', () => {
   });
 
   it('tracks source changes in the content data', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     media.source = { playbackId: 'abc123' };
     media.source = { playbackId: 'def456' };
 
@@ -95,21 +95,21 @@ describe('MuxMedia', () => {
   });
 
   it('has no content data without a playback id', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     media.source = { src: 'https://example.com/stream.m3u8' };
 
     expect(media.contentData).toEqual({});
   });
 
   it('has no content data for signed playback without image tokens', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     media.source = { playbackId: 'abc123', playback: { token: 'signed-playback-token' } };
 
     expect(media.contentData).toEqual({});
   });
 
   it('fires sourcechange when source is set', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     const onSourceChange = vi.fn();
     media.addEventListener('sourcechange', onSourceChange);
 
@@ -119,7 +119,7 @@ describe('MuxMedia', () => {
   });
 
   it('ignores the same source object', () => {
-    const media = new MuxMedia();
+    const media = new MuxVideoMedia();
     const source = { playbackId: 'abc123' };
     media.source = source;
 
@@ -133,6 +133,6 @@ describe('MuxMedia', () => {
   it('points unplayable-source copy at the hls.js-backed Media', () => {
     // The static exists for exactly this: SPF plays neither MPEG-TS nor DRM, and
     // the hls.js-backed Mux Media plays both.
-    expect(MuxMedia.alternativeMediaSuggestion).toContain('@videojs/media/dom/mux');
+    expect(MuxVideoMedia.alternativeMediaSuggestion).toContain('@videojs/media/dom/mux');
   });
 });
