+++ packages/spf/tsdown.config.ts
@@ -10,8 +10,8 @@ const createConfig = (mode: PackageBuildMode): UserConfig => ({
     hls: 'src/playback/engines/hls/index.ts',
     'media-tracks': 'src/media/media-tracks/index.ts',
     'background-video': 'src/playback/adapters/background-video/index.ts',
-    mux: 'src/playback/adapters/mux/index.ts',
-    'mux-audio-only': 'src/playback/adapters/mux-audio-only/index.ts',
+    'mux-audio': 'src/playback/adapters/mux-audio/index.ts',
+    'mux-video': 'src/playback/adapters/mux-video/index.ts',
     'simple-hls': 'src/playback/adapters/simple-hls/index.ts',
     'simple-hls-audio-only': 'src/playback/adapters/simple-hls-audio-only/index.ts',
   },
